+++ changelog.md
@@ -5,6 +5,7 @@
 ### Added
 * `[phonebook]` NoName contacts at the end of contacts list.
 * `[calllog]` Added new queries to the DB
+* `[audio]` Add storing and updating audio settings in database.
 
 ### Changed